refactor(ZNTA-1835): fix relative imports to work from new file locations
