+++ server/zanata-frontend/src/frontend/app/editor/components/DocsDropdown/index.js
@@ -1,5 +1,5 @@
 import Dropdown from '../Dropdown'
-import { Icon } from '../../components'
+import { Icon } from '../../../components'
 import React, { PropTypes } from 'react'
 import { Row } from 'react-bootstrap'
