+++ server/zanata-frontend/src/frontend/app/editor/components/IconButton/index.js
@@ -1,5 +1,5 @@
 import Button from '../Button'
-import { Icon } from '../../components'
+import { Icon } from '../../../components'
 import React, { PropTypes } from 'react'
 
 /**
