+++ server/zanata-frontend/src/frontend/app/editor/components/LanguagesDropdown/index.js
@@ -1,7 +1,7 @@
 import { values } from 'lodash'
 import { encode } from '../../utils/doc-id-util'
 import Dropdown from '../Dropdown'
-import { Icon } from '../../components'
+import { Icon } from '../../../components'
 import React, { PropTypes } from 'react'
 import { Row } from 'react-bootstrap'
